+++ NEWS.md
@@ -1,4 +1,9 @@
-# cmor.tools 0.3.0.9000
+# cmor.tools 0.4.0
+
+* New `scale_colour_cmor()`, and friends provide standard colour schemes.
+
+* Preliminary new `theme_cmor()` will provide a standard plot theme. Currently
+  only calls `ggplot2::theme_classic()`.
 
 # cmor.tools 0.3.0